refactor(unixdatepicker): clean up code and publish new UnixDatepicker component
Deleted all old versions of UnixDatepicker to keep the newest. Updated the project to build and
publish new component. Updated packages and now using newest version in demo.

closes #57
