--- demo/src/app/unix-datepicker-2/unix-datepicker-2.component.cy.ts
@@ -1,7 +1,0 @@
-import { UnixDatepicker2Component } from './unix-datepicker-2.component'
-
-describe('UnixDatepicker2Component', () => {
-  it('should mount', () => {
-    cy.mount(UnixDatepicker2Component)
-  })
-})
