--- demo/src/app/unix-datepicker-2/unix-datepicker-2.component.html
@@ -1,1 +1,0 @@
-<p>unix-datepicker-2 works!</p>
