--- demo/src/app/unix-datepicker-3/unix-datepicker-3.component.cy.ts
@@ -1,7 +1,0 @@
-import { UnixDatepicker3Component } from './unix-datepicker-3.component'
-
-describe('UnixDatepicker3Component', () => {
-  it('should mount', () => {
-    cy.mount(UnixDatepicker3Component)
-  })
-})
