--- demo/src/app/unix-datepicker-3/unix-datepicker-3.component.html
@@ -1,7 +1,0 @@
-<input #inputRef
-  [formControl]="internalControl"
-  [placeholder]="placeholder"
-  type="text"
-  inputmode="numeric"
-  (focus)="onFocus()"
-  (blur)="_onTouched(); onBlur()" />
