--- demo/src/app/unix-datepicker-3/unix-datepicker-3.component.scss
@@ -1,18 +1,0 @@
-input {
-  border: none;
-  background: none;
-  padding: 0;
-  outline: none;
-  font: inherit;
-  color: currentColor;
-  width: 100%;
-}
-
-:host {
-  opacity: 0;
-  transition: opacity 200ms;
-}
-
-:host.floating {
-  opacity: 1;
-}
